overhaul simulator by integrating historical and live generation modes
Strip historicalData.cpp to be purely responsible for CSV parsing. Migrate main.cpp to use purpose-built JSON parsing library instead of relying on fragile regex patterns. Change simulator types to be consistent with def.hpp. Remove unneeded or placeholder functions. Fix various runtime errors within engine.

+++ core/src/simulator/persistence_queue.hpp
@@ -25,8 +25,8 @@ struct OrderCancelledEvent {
 };
 
 struct BalanceEvent {
-    int balance;
-    int timestamp;
+    i64 balance;
+    u32 timestamp;
 };
 
 struct PositionSnapshotEvent {
